# perf: Intersection: Avoid callid AllIds inside inner loop